requirements.txt
# Core dependencies for Lab Assistant AI
# TODO: Add specific versions and test compatibility

# Audio processing
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0

# Speech-to-text
faster-whisper>=0.10.0
# Alternative: vosk>=0.3.45

# NLP and LLM
openai>=1.0.0  # For API-based LLM
# Alternative: transformers>=4.30.0  # For local LLM
# langchain>=0.1.0  # Optional: for agent framework

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0

# UI (optional)
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Memory/Vector store (for future use)
# faiss-cpu>=1.7.4
# chromadb>=0.4.0

